# Use LINDEX for consistent performance when calculating queue latency